% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{addArrayPositionNoRotate}
\alias{addArrayPositionNoRotate}
\title{Add the array row, column and index to a GAL file}
\usage{
addArrayPositionNoRotate(df, gridsPerRow = 4)
}
\description{
Add the array row, column and index to a GAL file
}

