% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addBarcodes}
\alias{addBarcodes}
\title{Add in the barcodes the MEPs came from}
\usage{
addBarcodes(dt3, dt4)
}
\arguments{
\item{dt3}{A spot level datatable with MEP_Drug and Barcode columns}

\item{dt4}{A MEP level datatable with a MEP_Drug column}
}
\value{
The input MEP level datatable with a Barcode column added
}
\description{
Add a barcode column that has one or more barcodes of the data's well plates
}
