% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{addMedianNormalizedInts}
\alias{addMedianNormalizedInts}
\title{Add columns of median normalized intensity data}
\usage{
addMedianNormalizedInts(DT, span = 0.1, ...)
}
\description{
Add columns of median normalized intensity data
}

