% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{addNormalizedInts}
\alias{addNormalizedInts}
\title{Add columns of loess normalized SpotCellCount and intensity data}
\usage{
addNormalizedInts(DT, span = 0.1, ...)
}
\description{
Add columns of loess normalized SpotCellCount and intensity data
}

