% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEPLINCSFunctions.R
\name{addOmeroIDs}
\alias{addOmeroIDs}
\title{Add the URLs for the Omero IDs on the OHSU Omero server}
\usage{
addOmeroIDs(dt)
}
\arguments{
\item{dt}{A datatable with an ImageID column}
}
\value{
the dt datatable with OmeroDetailURL, OmeroThumbnailURL and OmeroImageURL columns added
}
\description{
Add the URLs for the Omero IDs on the OHSU Omero server
}
