% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{addPositionParms}
\alias{addPositionParms}
\title{\code{addPositionParms} Add positional derived parameters to Cell data}
\usage{
addPositionParms(DT, outerThresh = 0.2, wedges = 18, sparseThresh = 0.8)
}
\arguments{
\item{DT}{A cell level datatable with plate-level x,y coordinates}

\item{outerThresh}{A numeric value to threshold the distance from the local origin to the cell to classify cells as Outer cells.}

\item{wedges}{A numeric value for the number of wedge-shpaed bins for grouping the theta values.}
}
\value{
A cell level datatable with new columns of XLocal, YLocal, RadialPostion, Theta, Outer, Density, Sparse and Perimeter values.
}
\description{
\code{addPositionParms} Add positional derived parameters to Cell data
}

