% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_levels.R
\name{addSpotProportions}
\alias{addSpotProportions}
\title{Add the porportions of gated cells in the spot populations}
\usage{
addSpotProportions(dt)
}
\arguments{
\item{dt}{A datatable}
}
\value{
nothing is returned, the proportions are added to the input datatable
}
\description{
Add the porportions of gated cells in the spot populations
}
