% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{autoMMCluster}
\alias{autoMMCluster}
\title{Cluster using a Mixture Model}
\usage{
autoMMCluster(x, G = 2)
}
\arguments{
\item{x}{A numeric vector}

\item{G}{the number of groups or clusters}
}
\value{
The cluster assignments of the values in x using Mclust from the mclust package.
}
\description{
\code{autoMMCluster} uses a Mixture Model to assign clusters.
}
