% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boundedLog2}
\alias{boundedLog2}
\title{Utility function that returns log2 with lower non-infinite bound}
\usage{
boundedLog2(x)
}
\description{
Utility function that returns log2 with lower non-infinite bound
}
