% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boundedLogit}
\alias{boundedLogit}
\title{Utility function that returns logit with lower and upper non-infinite bound}
\usage{
boundedLogit(x)
}
\arguments{
\item{x}{Vector to be logit transformed}
}
\value{
logit values of x
#' @export
}
\description{
Utility function that returns logit with lower and upper non-infinite bound
}
