% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{calc2NProportion}
\alias{calc2NProportion}
\title{Calculate the proportion of cells that are classfied as having 2N DNA}
\usage{
calc2NProportion(x)
}
\arguments{
\item{x}{numeric vector of cycle states with values of 1 for 2n and 2 for 4N}
}
\value{
proportion of cells in x that are in 2N
}
\description{
Calculate the proportion of cells that are classfied as having 2N DNA
}
