% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{calcProportion}
\alias{calcProportion}
\title{Calculate the proportion of x that are TRUE}
\usage{
calcProportion(x)
}
\arguments{
\item{x}{A logical vector}
}
\value{
The proportion of X that are TRUE
}
\description{
Calculate the proportion of x that are TRUE
}
