% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_functions.R
\name{calcQAScore}
\alias{calcQAScore}
\title{Calculate a QA score for MEMAs}
\usage{
calcQAScore(DT, threshold, maxNrSpot = 700, value)
}
\arguments{
\item{DT}{A datatable with MEMA data}

\item{threshold}{A numeric vector of length 1}

\item{maxNrSpot}{A numeric vector of length 1 that is the maximum number of spots printed on the MEMA}

\item{value}{A character vector of length 1 that is the column name of the values that determine the QA score}
}
\value{
A single numeric value that is the proportion of spots below the threshold
}
\description{
Return a QA score based on the number of spots missing or below a threshold
}
