% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{calcResidual}
\alias{calcResidual}
\title{Calculate the residuals from the median of a vector of numeric values}
\usage{
calcResidual(x)
}
\description{
Calculate the residuals from the median of a vector of numeric values
}

