% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{calcTheta}
\alias{calcTheta}
\title{\code{calcTheta} Return the polar coordinate theta value}
\usage{
calcTheta(x, y)
}
\arguments{
\item{x}{The x numeric value in cartesian coordinates.}

\item{y}{The y numeric value in cartesian coordinates.}
}
\value{
The theta value of the X and y converted to polar coordinates.


Function from Roland on Stack Overflow
http://stackoverflow.com/questions/23018056/convert-cartesian-angles-to-polar-compass-cardinal-angles-in-r
}
\description{
\code{calcTheta} Return the polar coordinate theta value
}

