% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{calculateZPrimeDT}
\alias{calculateZPrimeDT}
\title{Calculate the Z prime factor in a datatable}
\usage{
calculateZPrimeDT(DT)
}
\description{
Calculate the Z prime factor in a datatable
}

