% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{cellNeighbors}
\alias{cellNeighbors}
\title{Count the number of neigbors around a nuclei}
\usage{
cellNeighbors(spot, radius = (max(spot$Nuclei_CP_AreaShape_Center_X) -
  min(spot$Nuclei_CP_AreaShape_Center_X))/5)
}
\arguments{
\item{spot}{A data.table with \code{spot$Nuclei_CP_AreaShape_Center_X} and 
\code{spot$Nuclei_CP_AreaShape_Center_Y} values for each cell.}

\item{radius}{The radius around the center of each cell used to count neighbor cells.}
}
\description{
Count the number of neigbors around a nuclei
}
