% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{cleanColumnNames}
\alias{cleanColumnNames}
\title{Change the periods in the column names to spaces}
\usage{
cleanColumnNames(dt)
}
\arguments{
\item{dt}{a datatable}
}
\value{
the same datable with spaces in place of periods.
}
\description{
\code{cleanColumnNames} substitues a space for any single period in the colmun names of a datatable. This is useful at the end of an anlaysis to have human readable display names.
}

