% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_signal_functions.R
\name{cleanCommonSignalDt}
\alias{cleanCommonSignalDt}
\title{Clean the common signal datatable}
\usage{
cleanCommonSignalDt(dt)
}
\arguments{
\item{dt}{A datatable}
}
\value{
The input datatable without low quality data and normalized features
}
\description{
Remove normalized signals and low quality data
}
