% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{cleanExcelECMps}
\alias{cleanExcelECMps}
\title{Remove details from ECM protein names}
\usage{
cleanExcelECMps(x)
}
\arguments{
\item{x}{A chracter vector of ECM protein names}
}
\value{
The shortened character vector of the ECM protein names
}
\description{
Remove all details after the first underscore. Highlight proteins that are paired with COL1
by repositioning COL1 to be last.
}
