% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{cleanLegacyIssues}
\alias{cleanLegacyIssues}
\title{Clean up issues that have been problems in previous pipelines}
\usage{
cleanLegacyIssues(dt)
}
\arguments{
\item{dt}{A datatable of MEMA data and metadata}
}
\value{
The same datatable with legacy issues corrected
}
\description{
Clean up issues that have been problems in previous pipelines
}
