% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reformatCV.R
\name{coerceALLECM}
\alias{coerceALLECM}
\title{Coerce the All ECM contents to standard format}
\usage{
coerceALLECM(df)
}
\arguments{
\item{df}{A dataframe read from a gal file}
}
\value{
A dataframe with the Name column reformatted to match the Controlled Vocabulary standard.
}
\description{
\code{coerceALLECM}
}

