% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_signal_functions.R
\name{combineSignals}
\alias{combineSignals}
\title{Combine the common signals into a datatable}
\usage{
combineSignals(dtL)
}
\arguments{
\item{dtL}{a list of 1 to 3 datatables}
}
\value{
a datatable with the columns that are common across all datatables in dtL
}
\description{
Combine the common signals into a datatable
}
