% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compressHA}
\alias{compressHA}
\title{Replace the long names for hyaluronic acid with abbreviations}
\usage{
compressHA(x)
}
\arguments{
\item{x}{a chanracter vector that may contain names to be abbreviated}
}
\value{
The same character vector with abbrevaitaions as needed. hyaluronic_acid_greater_than_500kDa becomes
HA>500kDa and hyaluronic_acid_less_than_500kDa becomes HA<500kDa.
}
\description{
Replace the long names for hyaluronic acid with abbreviations
}
