% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{convertColumnNames}
\alias{convertColumnNames}
\title{Convert column names in a data.table}
\usage{
convertColumnNames(DT)
}
\arguments{
\item{DT}{A data.table}
}
\value{
DT The same data.table with duplicated columns, invalid column name characters and trailing spaces removed.
}
\description{
Convert column names in a data.table
}
