% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEPLINCSFunctions.R
\name{create8WellHistograms}
\alias{create8WellHistograms}
\title{Create histograms of 8 well MEMA data}
\usage{
create8WellHistograms(DT, pr, prDisplay,
  binwidth = diff(quantile(DT[[pr]], probs = c(0, 0.98), na.rm =
  TRUE))/50, upperProb = 0.99, ncol = 4)
}
\arguments{
\item{DT}{A datatable with a pr column of values}

\item{pr}{The name of the column of values to be shown in the images}

\item{prDisplay}{A name for the legend title}

\item{binwidth}{Optional binwidth value for the histograms}

\item{upperProb}{Optional upper probablity to limit the display}

\item{ncol}{Optional number of columns to display in the ggplot facet}
}
\value{
The ggplot grob to be displayed
}
\description{
Create histograms of 8 well MEMA data
}
