% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEPLINCSFunctions.R
\name{create8WellPseudoImage}
\alias{create8WellPseudoImage}
\title{Create pseudeoimages of 8 well MEMA data}
\usage{
create8WellPseudoImage(DT, pr, prDisplay)
}
\arguments{
\item{DT}{A datatable with ArrayRow, ArrayColumn and pr columns}

\item{pr}{The name of the column of values to be shown in the images}

\item{prDisplay}{A name for the legend title}
}
\value{
The ggplot grob to be displayed
}
\description{
Create pseudeoimages of 8 well MEMA data
}
