% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_levels.R
\name{createl4}
\alias{createl4}
\title{Summarize spot level data to the MEP level}
\usage{
createl4(l3, seNames = NULL)
}
\arguments{
\item{l3}{The datatable of spot level data to be summarized}
}
\value{
A datatable of MEP level, median summarized data with standard error values and 
 metadata
}
\description{
Median summarize the spot level normalized values the most biologically 
interpretable raw data at each spot, calculate the standard errors and add
SE columns for all median summarized data
}
