% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{createRUVM}
\alias{createRUVM}
\title{Create an M matrix for the RUV normalization
that has one row for each unit to be normalized and
one column for each unique unit type
Each row will have a 1 to indicate the unit type
all other values will be 0}
\usage{
createRUVM(dt)
}
\description{
Create an M matrix for the RUV normalization
that has one row for each unit to be normalized and
one column for each unique unit type
Each row will have a 1 to indicate the unit type
all other values will be 0
}

