% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRUVM.R
\name{createRUVMGeneral}
\alias{createRUVMGeneral}
\title{Create an M matrix for the RUV normalization}
\usage{
createRUVMGeneral(dt, unitID = "BWL", uniqueID = "Ligand")
}
\arguments{
\item{dt}{The datatable to be normalized. There must be a SignalType column 
where a value of "Signal" denotes which rows should be included in the M matrix.}

\item{unitID}{The column name that identifies the names of the units to be normalized.
For example, this may have the value BW as the barcode and well can be combined to 
create unique identifiers for a unit of data.}

\item{uniqueID}{The column name that uniquely identifies the replicates. For example,
this may have a value of Ligand or LigandDrug depending on the experiment.}
}
\value{
A datatable with values of 1 and 0 that captures the structure of dt.
}
\description{
The M matrix holds the structure of the dataset in RUV normalization.
There is one row for each unit to be normalized and
one column for each unique unit type
Each row will have one 1 to indicate the unit type
all other values will be 0
}

