% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{extractNamedValues}
\alias{extractNamedValues}
\title{Extract the named value from a dataset}
\usage{
extractNamedValues(normedPlate, namedValue)
}
\description{
Extract the named value from a dataset
}

