% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{extractNormalizedCntrls}
\alias{extractNormalizedCntrls}
\title{Add columns of loess normalized intensity data}
\usage{
extractNormalizedCntrls(DT, span, ...)
}
\description{
Add columns of loess normalized intensity data
}

