% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{filterWCC}
\alias{filterWCC}
\title{remove any wells (and its replicates) that have fewer than wccThresh cells}
\usage{
filterWCC(DT, wccThresh)
}
\description{
remove any wells (and its replicates) that have fewer than wccThresh cells
}

