% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{findPerimeterCell}
\alias{findPerimeterCell}
\title{\code{findPerimeterCell} Determine the perimeter cell in wedge}
\usage{
findPerimeterCell(x)
}
\arguments{
\item{x}{A datatable or dataframe with a RadialPosition column}
}
\value{
A logical vector the length of x with a TRUE value for the Perimeter cell
}
\description{
\code{findPerimeterCell} Determine the perimeter cell in wedge
}
