% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-firing_rate_by_treatment.R
\name{firing_rate_by_treatment}
\alias{firing_rate_by_treatment}
\title{Dot/boxplot plot of average firing rate by treatment}
\usage{
firing_rate_by_treatment(firing, experiment_tag, plot_width = 6,
  plot_height = 6, output_base = "product/figures", verbose = TRUE)
}
\description{
Make a plot where
     the x-axis are the treatments and
     the y-axis is the average firing rate on the log scale
     show a box (and wisker) plot for each treatment summarizing across neurons in the experiment
     overlay points for each neuron jittering horizontally to allow them to be seen more easily
}
\details{
firing:
  data.frame with columns [neuron_index, timestamp, treatment, begin, end]

experiment_tag:
  identifier for the experiment used in the figure subtitle and output filename

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/figures/firing_rate_by_treatment_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
