% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{formatCP96Well}
\alias{formatCP96Well}
\title{Format 96 well plate raw data from the Cell Profiler Pipeline}
\usage{
formatCP96Well(dt, nrArrayRows, nrArrayColumns)
}
\arguments{
\item{dt}{data.table with a Spot column}

\item{nrArrayRows}{The number of rows in the printed array}

\item{nrArrayColumns}{The number of columns in the printed array}
}
\value{
A data.table with an alphanumeric Well column and reindexed Spot column
}
\description{
Reformats the metadata for one row of a 96 well plate to match the pipeline. 
The Spot column is converted to an index within each well and a label for the Well is added
to the output data.table
}
