% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{gateCells}
\alias{gateCells}
\title{Gate cells based on specific stains if they are in the dataset}
\usage{
gateCells(dt, synId = "syn10138929")
}
\arguments{
\item{dt}{A datatable that contains cell level data and metadata}
}
\description{
Gate cells based on specific stains if they are in the dataset
}
