% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gateOnlocalMinima}
\alias{gateOnlocalMinima}
\title{Gate a vector on a local minima}
\usage{
gateOnlocalMinima(x, probs = c(0.2, 0.8), ...)
}
\arguments{
\item{x}{numeric vector}

\item{probs}{interval limits on where to search for the minima}
}
\description{
Gate a vector on a local minima
}
