% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gateOnQuantile}
\alias{gateOnQuantile}
\title{Gate values at a specific quantile}
\usage{
gateOnQuantile(x, probs)
}
\arguments{
\item{x}{A vector of values}

\item{probsThe}{quantile for gating}
}
\value{
An integer vector the same length as x with values of 0 for the lower class and 1
as the upper class value
}
\description{
Gate values at a specific quantile
}
