% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Get barcodes from the Synapse Plate Tracker files}
\usage{
getBarcodes(studyName, synId = "syn8440875")
}
\arguments{
\item{studyName}{Character string of the study name}
}
\value{
Barcodes of the plates in the study
}
\description{
Get barcodes from the Synapse Plate Tracker files
}
