% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getCPData}
\alias{getCPData}
\title{Get the data from the Cell Profiler pipeline}
\usage{
getCPData(dataBWInfo, curatedOnly = TRUE,
  curatedCols = "ImageNumber|ObjectNumber|AreaShape|_MedianIntensity_|_IntegratedIntensity_|_Center_|_PA_|Texture",
  verbose = FALSE)
}
\description{
Get the data from the Cell Profiler pipeline
}
