% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getICData}
\alias{getICData}
\title{Read and convert INCell data to CP format}
\usage{
getICData(cellDataFilePaths, endPoint488, endPoint555, endPoint647,
  verbose = FALSE)
}
\arguments{
\item{cellDataFilePaths}{A character vector of the file paths to the cell level data}
}
\description{
Read and convert INCell data to CP format
}
