% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getLogMetadata}
\alias{getLogMetadata}
\title{Get deposition values from an Aushon log dta files}
\usage{
getLogMetadata(path, barcode)
}
\arguments{
\item{path}{path to the directory with barcode level subdirectories}

\item{barcode}{barcode or plate ID}
}
\value{
a datatable with an integer Depostion column
}
\description{
Get deposition values from an Aushon log dta files
}
