% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata from either An! or !An! files}
\usage{
getMetadata(barcode, path, useAnnotMetadata = TRUE)
}
\description{
Get metadata from either An! or !An! files
}
