% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{getOmeroIDs}
\alias{getOmeroIDs}
\title{Read in and merge  Omero URLs}
\usage{
getOmeroIDs(path)
}
\arguments{
\item{path}{The path to the file named barcode_imageIDs.tsv}
}
\value{
a datatable with WellIndex, ArrayRow, ArrayColumn and ImageID columns
}
\description{
Adds Omero image IDs based on the WellName values
}
