% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpotLevelData}
\alias{getSpotLevelData}
\title{Get the spot level data for a study}
\usage{
getSpotLevelData(paths)
}
\arguments{
\item{paths}{The paths to the spot level files}
}
\value{
A datatable with the annotated data for all plates in the study. Any data for fiducials and 
blank spots is filtered out.
}
\description{
Get the spot level data for a study
}
