% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getSpotMetadata}
\alias{getSpotMetadata}
\title{Read the spot metadata from a gal file}
\usage{
getSpotMetadata(path, barcode)
}
\arguments{
\item{path}{path to the directory with barcode level subdirectories}

\item{barcode}{barcode or plate ID}
}
\value{
a datatable with the spot level metadata
}
\description{
Read the spot metadata from a gal file
}
