% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getWellMetadata}
\alias{getWellMetadata}
\title{Read the well metadata from a multi-sheet Excel file}
\usage{
getWellMetadata(fn)
}
\arguments{
\item{fn}{full path and file name for the well metadata file}
}
\value{
a datatable with the well level metadata
}
\description{
Read the well metadata from a multi-sheet Excel file
}
