% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{kmeansCluster}
\alias{kmeansCluster}
\title{Cluster using kmeans}
\usage{
kmeansCluster(x, centers = 2)
}
\arguments{
\item{x}{A numeric vector to be clustered}

\item{centers}{The number of centers or clusters to find.}
}
\value{
The cluster assignments for x using the base kmeans command.
}
\description{
\code{kmeansCluster} is a wrapper function for perfoming kmeans clustering
}

