% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{labelOuterCells}
\alias{labelOuterCells}
\title{\code{labelOuterCells} Determine if a cell is in Outer cell in the spot}
\usage{
labelOuterCells(x, thresh = 0.75)
}
\arguments{
\item{x}{A numeric vector of RadialPositions}

\item{thresh}{A numeric value to threshold x. The returned logical will be TRUE for cells with x values greater than thresh.}
}
\value{
A logical vector the length of x with a TRUE value for the Outer cells
}
\description{
\code{labelOuterCells} Determine if a cell is in Outer cell in the spot
}

