% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Load firing data for an experiment}
\usage{
load_experiment(units_fname, treatments_fname, experiment_tag = NULL,
  save_to_disk = TRUE, verbose = TRUE)
}
\description{
units_fname:
  a file should be a .mat file exported from
  the Margolis lab multi-electrod expirmental matlab processing pipeline
  containing the following information for each neuron that was detected
}
\details{
1) the time steps (in seconds) when it fired during the experiment
    2) the waveform of the cluster center across all firings

treatments_fname:
  a .csv file with columns [treatment,begin,end] for each treatment in the experiment
  to help detect problems, an error is returned if the treatments are not disjoint and given chronologically

experiment_tag:
  identifier for the experiment, set in the return data structure and path to save to disk
  if null (default), then use treatments_fname %>% basename %>% str_replace(".mat$", "")

save_to_disk:
   if TRUE then the return value is saved to
   paste0("intermediate_data/", <experiment_tag>, ".Rdata")

returns:
   an experiment S3 class with the following elements
     tag: <experiment_tag>
     treatments: tibble::tibble with columns [treatment, begin, end]
     firings: tibble::tibble with columns [nuron_index, time_step, treatment, begin, end] and a row for each detected firing
     waveform: tibble::tibble with columns [neuron_index, time_step, voltage] for each neuron
     class attribute: mema_experiment
}
