% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{loessModel}
\alias{loessModel}
\title{loessModel Create a median normalized loess model of an array}
\usage{
loessModel(data, value, span)
}
\arguments{
\item{data}{A dataframe with ArrayRow, ArrayColumn and signal intensity columns}

\item{value}{The column name of the signal intensity column}

\item{span}{The span value passed to loess. Values between 0 and 1 determine the
proportion of the population to be included in the loess neighborhood.}
}
\value{
a vector of median normalized loess values of the signal
}
\description{
loessModel Create a median normalized loess model of an array
}

