% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{loessNorm}
\alias{loessNorm}
\title{Loess normalize an array using the spatial residuals}
\usage{
loessNorm(Value, Residual, ArrayRow, ArrayColumn)
}
\description{
Loess normalize an array using the spatial residuals
}
