% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV_normalize_functions.R,
%   R/normalize_functions.R
\name{loessNormArray}
\alias{loessNormArray}
\title{Loess normalize values within an array}
\usage{
loessNormArray(dt)

loessNormArray(dt)
}
\description{
Loess normalize values within an array

Loess normalize values within an array
}
