% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{makeValidColumnNames}
\alias{makeValidColumnNames}
\title{Make valid column names in a data.table}
\usage{
makeValidColumnNames(DT)
}
\arguments{
\item{DT}{A data.table}
}
\value{
DT The same data.table with intenal paces in the column names converted to . Any  trailing spaces are deleted.
}
\description{
Make valid column names in a data.table
}

