% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{medianNorm}
\alias{medianNorm}
\title{Median normalize a vector}
\usage{
medianNorm(DT, value)
}
\arguments{
\item{DT}{A datatable with a column \code{value}}

\item{value}{The name of the column in DT to be normalized}
}
\value{
A numeric vector that is \code{DT[[value]]} dvided by its median
}
\description{
Median normalize a vector
}

