% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/population_annotate_functions.R
\name{melt8Well}
\alias{melt8Well}
\title{Melt raw population dataset to combine wells}
\usage{
melt8Well(DT, swap = TRUE)
}
\arguments{
\item{DT}{a data.table with columns of population intensity values in columns named by barcode, well, wavelength and data type(Raw/Net/Background)}

\item{swap}{a logical on whether the data for the A and B rows should be swapped. This must be yes for 8 well data coming from the Array Pro software.}
}
\value{
A data.table in long format with columns of intensity readings for #' each data type and wavelength, barcode and well.
}
\description{
\code{melt8Well} Converts a Tecan ArrayPro data file to a data.table with properly labeled columns.
}
\section{Usage}{
 This function reformats Array Pro results files that
come from analyzing images from multiple channels, wells and plates.
The column names are encoded as follows:
Datatype is at the beginning of the column name as Raw, Net or
Background. Barcode follows the curly brace symbol. Wavelength follows
 the - symbol after the barcode. Well follows the - symbol after the
 wavelength. The A and B well labels are swapped to align the Tecan
 labeling to the actual well names.
}

