% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{mergeCytoMain}
\alias{mergeCytoMain}
\title{Merge the Cyto and Main data files}
\usage{
mergeCytoMain(cydt, mdt)
}
\arguments{
\item{cydt}{The datatable of the cytoplasmic data.}

\item{mdt}{The datatable of the main data.}
}
\value{
A datatable of the merged data.
}
\description{
\code{mergeCytoMain} merges the Olympus scan^R cytoplasmic and main datatables. The Object.ID in the main datatable is merged with Parent.Object.ID..MO in the cyto datatable.
}

