% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{mergeSpot8WellMetadata}
\alias{mergeSpot8WellMetadata}
\title{Create a single datatable for all wells in an 8 well plate using!An! metadata}
\usage{
mergeSpot8WellMetadata(spotMetadata, wellMetadata)
}
\arguments{
\item{spotMetadata}{A datatable of spot level metadata}

\item{wellMetadata}{A datatable of well level metadata}
}
\value{
A spot level datatable that has the well level metadata properly merged
}
\description{
8 well MEMAs are printed with the B row rotated 180 degrees from the A row. This routine rotates
the B rows and creates. The Spot values are indices in the physical coordinate system.
PrintSpot values are indices in the rotated printing coordinate system.
}
