% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{mergeSpot96WellMetadata}
\alias{mergeSpot96WellMetadata}
\title{Merge the spot well level metadata for a 96 well plate}
\usage{
mergeSpot96WellMetadata(spotMetadata, wellMetadata)
}
\arguments{
\item{spotMetadata}{A datatable of spot level metadata}

\item{wellMetadata}{A datatable of well level metadata}
}
\value{
A spot level datatable that has the well level metadata properly merged
}
\description{
This function assumes a 12 pin printhead in 6 row and 2 column configuration.
The printhead is oriented 90 ccw from the well plate. The 12 well pattern is
duplicated 8 times in the well plate to print 8 identical sets of 12 arrays
}
