% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{mergeSpotMetadata}
\alias{mergeSpotMetadata}
\title{merge the data and metadata on the well index column
\code{mergeSpotMetadata} merges the well metadata with data on the well value.}
\usage{
mergeSpotMetadata(dt, mdf)
}
\arguments{
\item{dt}{A datatable of data values}

\item{mdf}{A datatable of metadata}
}
\value{
A datatable of annotated data that is merged on the Well values.
}
\description{
merge the data and metadata on the well index column
\code{mergeSpotMetadata} merges the well metadata with data on the well value.
}

