% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV_normalize_functions.R
\name{norm_RR_RUVResiduals}
\alias{norm_RR_RUVResiduals}
\title{Apply RUV Normalization to the signals in a MEMA dataset}
\usage{
norm_RR_RUVResiduals(dt, k)
}
\arguments{
\item{dt}{A datatable with data and metadata to be normalzed. There must be CellLine, Barcode, Well, Ligand, Drug and ECMp metadata columns}

\item{k}{The number of factors to be removed in the RUV normalization}
}
\description{
Apply RUV Normalization to the signals in a MEMA dataset
}
