% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normProfToCol1}
\alias{normProfToCol1}
\title{Normalize the proliferation ratio signal to the collagen 1 values}
\usage{
normProfToCol1(x)
}
\arguments{
\item{x}{a dataframe or datatable with columns names ProliferatioRatio
and ShortName. ShortName must include at least one entry of COL1 or COL I.}
}
\value{
The input dataframe of datatable with a normedProliferation column that has the ProliferationRatio values divided by the median collagen
1 proliferation value
}
\description{
Normalize the proliferation ratio signal to the collagen 1 values
}

