% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normRUVLoessResiduals}
\alias{normRUVLoessResiduals}
\title{Apply RUV and Loess Normalization to the signals in a dataset}
\usage{
normRUVLoessResiduals(dt, k)
}
\description{
Apply RUV and Loess Normalization to the signals in a dataset
}

