% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeCSx2.R
\name{normRUVLoessResidualsCS}
\alias{normRUVLoessResidualsCS}
\title{Apply RUV and Loess Normalization to the common signals in a dataset}
\usage{
normRUVLoessResidualsCS(dt, k)
}
\description{
Apply RUV and Loess Normalization to the common signals in a dataset
}
