% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeCSx2.R
\name{normRUVLoessResidualsCSx2}
\alias{normRUVLoessResidualsCSx2}
\title{Apply RUV and Loess Normalization to the common signals in a dataset}
\usage{
normRUVLoessResidualsCSx2(dt, k)
}
\description{
Apply RUV and Loess Normalization to the common signals in a dataset
}
