% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normRUVLoessResidualsDisplay}
\alias{normRUVLoessResidualsDisplay}
\title{Apply RUV and Loess Normalization to the signals in a dataset}
\usage{
normRUVLoessResidualsDisplay(dt, k)
}
\description{
Apply RUV and Loess Normalization to the signals in a dataset
}

