% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV_normalize_functions.R
\name{normRUVResiduals}
\alias{normRUVResiduals}
\title{Apply RUV Normalization to the signals in a MEMA dataset}
\usage{
normRUVResiduals(dt, k)
}
\arguments{
\item{dt}{A datatable with data and metadata to be normalzed. There must be CellLine, Barcode, Well, Ligand, Drug and ECMp metadata columns}

\item{k}{The number of factors to be removed in the RUV normalization}
}
\description{
Apply RUV Normalization to the signals in a MEMA dataset
}
