% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normRZSWellsWithinPlate}
\alias{normRZSWellsWithinPlate}
\title{RZS Normalize a Column of Data}
\usage{
normRZSWellsWithinPlate(DT, value, baseECM, baseL)
}
\arguments{
\item{DT}{A datatable with value, baseECM and baseL, ECMp and
Ligand columns}

\item{value}{A single column name of the value to be normalized}

\item{baseECM}{A single character string or a regular expression that selects
the ECM(s) that are used as the base for normalization.}

\item{baseL}{A single character string or a regular expression that selects
the ligand used as the base for normalization.}
}
\value{
a vector of RZS normalized values
}
\description{
\code{normRZSWellsWithinPlate} normalizes all elements of DT[[value]] by
subtracting the median of DT[[value]] of all baseECM spots in the
baseL wells, then divides the result by the MAD*1.48 of all baseECM spots in
the baseL wells
}

