% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{normWellsWithinPlate}
\alias{normWellsWithinPlate}
\title{Normalize to a base MEP}
\usage{
normWellsWithinPlate(DT, value, baseECM, baseGF)
}
\arguments{
\item{DT}{A \code{data.table} that includes a numeric value column to be
normalized, a \code{ShortName} column that has the printed ECM names and a
\code{Growth.Factors} column that has the growth factor names.}

\item{value}{The name of the column of values to be normalized}

\item{baseECM}{The name of the printed ECM to be normalized against}

\item{baseGF}{The name of the soluble growth factor to be normalized against}
}
\value{
A numeric vector of the normalized values
}
\description{
Normalizes one channel of values for all MEPs in a multi-well plate to one
base MEP.
}
\section{Details}{
 \code{normWellsWithinPlate} normalizes the value column of
  all MEPs by subtracting the median value of the replicates of the MEP that
  is the pairing of baseECM  with baseGF.
}

