% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{numericMedian}
\alias{numericMedian}
\title{Return the median of a vector as a numeric value}
\usage{
numericMedian(x)
}
\arguments{
\item{x}{integer or double vector}
}
\value{
The median of x as a numeric value
}
\description{
\code{numericMedian} is a helper function for use within data.table that ensure the all medians are returned as numeric instead of numeric or integer values.
}
