% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gal_parse_functions.R
\name{parse4wellvalidationContents}
\alias{parse4wellvalidationContents}
\title{Parse 4wellvalidationContents for ECM, Glycerol, Triton, EDTA and Buffer values}
\usage{
parse4wellvalidationContents(dt)
}
\arguments{
\item{dt}{a datatable with a Name column to be parsed}
}
\value{
The input datatable with new columns of ECM, Glycerol, Triton, EDTA and Buffer.
}
\description{
\code{parse4wellvalidationContents} returns a datatable with new columns based on the Name column.
}
\section{Usage}{
 This functions assumes a Name column exists and uses an underscore as field separators.
}

