% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gal_parse_functions.R
\name{parseBufferWinnerContents}
\alias{parseBufferWinnerContents}
\title{Parse BufferWinners names for ECM, Ligand and Concentration values}
\usage{
parseBufferWinnerContents(dt)
}
\arguments{
\item{dt}{a datatable with a Name column to be parsed}
}
\value{
The input datatable with new columns of Condition1, Condition2, Condition3 and Condition4.
}
\description{
\code{parseBufferWinnerContents} returns a datatable with new columns based on the Name column.
}
\section{Usage}{
 This functions assumes a Name column exists uses an underscore as field separators.
}

