% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gal_parse_functions.R
\name{parseContents}
\alias{parseContents}
\title{Parse names for ECM, Ligand and Concentration values}
\usage{
parseContents(dt)
}
\arguments{
\item{dt}{a datatable with a Name column to be parsed}
}
\value{
The input datatable with new columns for the ECM, Ligand and
 Concentration values.
}
\description{
\code{parseContents} returns a datatable with new columns for the ECM, ligand
and concentration at each spot.
}
\section{Usage}{
 This functions assumes a Name column exists and that its
 format is ECM_Ligand_Concentration. If there are no underscores, then the
 Name is interpreted as an ECM, a NULL string is assigned to the Ligand and a
 concentration value of 0 is assigned to Concentration. If there are 2 words
 separated by an underscore, the first is the ECM and the second is the
 Ligand. The concentration is assigned a value of 0. If there are 3 words,
 the first is assigned to the concentration.

 In the special case of the AF 488 fiducial, the Ligand is assigned the value
 AF 488.
}

