% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gal_parse_functions.R
\name{parseCVContents}
\alias{parseCVContents}
\title{Parse controlled vocabulary data for ECM, Ligand and ConcentrationRank values}
\usage{
parseCVContents(dt)
}
\arguments{
\item{dt}{A dataframe or datatable with a Name column to be parsed.}
}
\value{
The input datatable with a new column of the ECM names pasted
 togethers with an underscore.
}
\description{
\code{parseCVContents} returns a datatable with new columns based on the Name
column.
}
\section{Usage}{
 This functions assumes a Name column exists and uses a dash
 symbol - as a field separator between each protein and an underscore between
 a protein's name and its Uniprot ID.
}

