% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gal_parse_functions.R
\name{parseSimulatedContents}
\alias{parseSimulatedContents}
\title{Parse simulated data for ECM, Ligand and ConcentrationRank values}
\usage{
parseSimulatedContents(dt)
}
\arguments{
\item{dt}{a datatable with a Name column to be parsed}
}
\value{
The input datatable with new columns of ECM, Ligand and ConcentrationRank.
}
\description{
\code{parseSimulatedContents} returns a datatable with new columns based on
the Name column.
}
\section{Usage}{
 This functions assumes a Name column exists and uses an
 underscore as field separators. The format is assumed to be
 ECM_Ligand_ConcentrationRank
}

