% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_firing_density_by_neuron.R
\name{plot_firing_density_by_neuron}
\alias{plot_firing_density_by_neuron}
\title{Firing smoothed per-neuron firing density by neuron across the experiment}
\usage{
plot_firing_density_by_neuron(
  experiment,
  plot_width = 10,
  plot_height = 4,
  output_base = "product/plots",
  verbose = TRUE
)
}
\description{
heat-map average firing rate
    x-axis experiment in seconds
    y-axis neuron index
    cell KDE smoothe estimate of the firing rate
}
\details{
experiment:
  experiment dataset loaded with mema::load_experiment(...)

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/firing_rate_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
