% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_firing_qqplot_by_treatment.R
\name{plot_firing_qqplot_by_treatment}
\alias{plot_firing_qqplot_by_treatment}
\title{Firing QQ-plot by treatment}
\usage{
plot_firing_qqplot_by_treatment(
  experiment,
  highlight_units = NULL,
  plot_width = 7,
  plot_height = 4,
  output_base = "product/plots",
  verbose = TRUE
)
}
\description{
Lattice plot for each treatment, where in each cell
     the x-axis is percent of the treatment (by time)
     the y-axis is percent of the total number of counts observed
     a line for each neuron in the experiment with
        the left endpoint anchored at the lower left
        the right endpoint anchored at the upper right
}
\details{
If the counts occure independently through time as is assumed in a poisson process model
 then for each neuron, the line be diagonal (following the guide blue line)
   If there is some lag to reach a stead state then the lines will look like they have a kink in it
   if there is bursty or irregular spacing, then the spead off the diagonal will be substantial but symmetric

experiment:
  experiment dataset loaded with mema::load_experiment(...)

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/firing_qqplot_by_treatment_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
