% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_firing_rate.R,
%   R/plot_firing_rate_by_neuron.R
\name{plot_firing_rate_by_neuron}
\alias{plot_firing_rate_by_neuron}
\title{Firing smoothed per-neuron firing rate across the experiment}
\usage{
plot_firing_rate_by_neuron(experiment, plot_width = 7, plot_height = 4,
  output_base = "product/plots", verbose = TRUE)

plot_firing_rate_by_neuron(experiment, plot_width = 7, plot_height = 4,
  output_base = "product/plots", verbose = TRUE)
}
\description{
heat-map average firing rate
    x-axis experiment in seconds
    y-axis neuron index
    cell KDE smoothe estimate of the firing rate

histogram of average firing rate
    x-axis firing rate on the log-scale
    y-axis number of neurons in the histogram bin
}
\details{
if there is spread in the average firing rate then this variation should be
 accounted for in a model of effects of the treatment on the average firing rate

experiment:
  experiment dataset loaded with mema::load_experiment(...)

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/firing_rate_by_neuron_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator

if there is spread in the average firing rate then this variation should be
 accounted for in a model of effects of the treatment on the average firing rate

experiment:
  experiment dataset loaded with mema::load_experiment(...)

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/firing_rate_by_neuron_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
