% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_conditional_effects.R
\name{plot_model_conditional_effects}
\alias{plot_model_conditional_effects}
\title{Plot conditional effects for model}
\usage{
plot_model_conditional_effects(
  model,
  model_tag = NULL,
  effect = "treatment",
  plot_width = 6,
  plot_height = 6,
  output_base = "product/plots",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{\code{brmsfit} object}

\item{model_tag}{\code{character} tag for the model (Default: \code{NULL})}

\item{plot_width}{\code{numeric} dimensions of the output plot width (Default: \code{6})}

\item{plot_height}{\code{numeric} dimensions of the output plot height (Default: \code{6})}

\item{output_base}{\code{character} path to saveplots (Default: \code{"product/plots"})}

\item{verbose}{\code{logical} verbose output (Default: FALSE)}

\item{...}{additional arguments passed to \code{brms::conditional_effects}}

\item{effects}{\code{character} main effects to plot (see \code{brms::conditional_effects})}
}
\value{
ggplot2 object
  Saves the result to product/plots/firing_rate_by_neuron_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
\description{
Plot conditional effects for model
}
