% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_marginal_effects.R
\name{plot_model_marginal_effects}
\alias{plot_model_marginal_effects}
\title{plot_width/plot_height:
  dimensions of the output plot}
\usage{
plot_model_marginal_effects(
  model_fit,
  model_tag,
  plot_width = 6,
  plot_height = 6,
  output_base = "product/plots",
  verbose = TRUE,
  ...
)
}
\description{
returns:
  the ggplot2
  Saves the result to product/plots/firing_rate_by_neuron_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
