% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unit_repsonse_correlation_by_treatment.R
\name{plot_unit_response_by_treatment}
\alias{plot_unit_response_by_treatment}
\title{Unit Response Correlation by Treatment}
\usage{
plot_unit_response_by_treatment(
  experiment,
  highlight_units = NULL,
  plot_width = 10,
  plot_height = 10,
  output_base = "product/plots",
  verbose = TRUE
)
}
\description{
Grid of treatment vs. treatment plots, in each cell
     scatter plot for each neuronal unit
       the x-axis is average firing rate for treatment of column
       the y-axis is average firing rate for treatment of row
}
\details{
experiment:
  experiment dataset loaded with mema::load_experiment(...)

highlight_units:
  which units to highlight in the plot

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/unit_response_correlation_by_treatment_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
