% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waveform_correlation_matrix.R
\name{plot_waveform_correlation_matrix}
\alias{plot_waveform_correlation_matrix}
\title{plot Correlation matrix of waveforms}
\usage{
plot_waveform_correlation_matrix(
  experiment,
  plot_width = 10,
  plot_height = 10,
  output_base = "product/plots",
  verbose = TRUE
)
}
\description{
Plot a correlation matrix as a heatmap between all waveforms in the experiment
}
\details{
experiment:
  experiment dataset loaded with mema::load_experiment(...)

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/waveform_correlation_matrix_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
