% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waveform_lattice.R
\name{plot_waveform_lattice}
\alias{plot_waveform_lattice}
\title{Simple lattice plot of the waveforms}
\usage{
plot_waveform_lattice(
  experiment,
  plot_width = 10,
  plot_height = 10,
  output_base = "product/plots",
  verbose = TRUE
)
}
\description{
Plot a grid of plots, with x-axis in microseconds and y-axis in voltage
  One for each waveform in the waveform data.frame
}
\details{
experiment:
  experiment dataset loaded with mema::load_experiment(...)

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/plots/waveform_lattice_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
