% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{plotLEHmap}
\alias{plotLEHmap}
\title{Plot a heatmap with ligand rows and ECM protein columns}
\usage{
plotLEHmap(dt, hclustfun = NULL, fill, titleExpression, limits,
  xAxisSize = 0.6, yAxisSize = 0.5)
}
\arguments{
\item{dt}{Datatable with Ligand, ECMp and a column named by the fill value}

\item{hclustfun}{Optional clustering function to overide heatmap.2 default}

\item{fill}{Name of column with numeric values for coloring the heatmap}

\item{titleExpression}{Title for heatmap}

\item{limits}{Length 2 integer or numeric vector used to squish the colors of extreme values}

\item{xAxisSize}{Optional scale factor for x axis text}

\item{yAxisSize}{Optional scale factor for y axis text}
}
\value{
none called for its side effect of plotting the heatmap
}
\description{
Plot a heatmap with ligand rows and ECM protein columns
}
