% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEPLINCSFunctions.R
\name{plotSCCHeatmapsQAHistograms}
\alias{plotSCCHeatmapsQAHistograms}
\title{Print pseudoimages and histograms of the spot cell count values}
\usage{
plotSCCHeatmapsQAHistograms(l3, barcodes, lthresh)
}
\arguments{
\item{l3}{A datatable with Barcode, ECMp, Spot_PA_SpotCellCount, Well_Ligand and QA_score columns}

\item{barcodes}{The barcodes of the data to be displayed}

\item{lthresh}{The value used to gate the poor quuality spots.}
}
\value{
none This function is called to print the images
}
\description{
Print pseudoimages and histograms of the spot cell count values
}
