% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{popDT}
\alias{popDT}
\title{Annotated data from a 4 plate experiment}
\format{A data table with 39 columns and 22,272 rows of data and metadata.}
\source{
Tecan LS Reloaded
}
\usage{
popDT
}
\description{
Three channels of annotated population level data from 32 MEMAs.
}
\keyword{datasets}

