% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_signal_functions.R
\name{preprocessCommonSignalsLevel3}
\alias{preprocessCommonSignalsLevel3}
\title{Normalize the common signals across studies}
\usage{
preprocessCommonSignalsLevel3(paths, k = 256L, verbose = FALSE)
}
\arguments{
\item{paths}{A list of paths to level 2 data, each element is a different study.}

\item{k}{The number of factors to use in the RUV normalization}

\item{verbose}{A logical for progress messages}
}
\value{
A datatable of raw and normalized spot level data with its metadata
}
\description{
Normalize the common signals across studies
}
