% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{processan2omero}
\alias{processan2omero}
\title{Process the metadata in an an2omero file}
\usage{
processan2omero(fileName)
}
\arguments{
\item{fileName}{The full name with path for the an2omero file}
}
\value{
A datatable where each row is a unique spot in one well plate
}
\description{
Process the metadata in an an2omero file
}
