% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QA_functions.R
\name{randomizePositions}
\alias{randomizePositions}
\title{Randomize a vector of values}
\usage{
randomizePositions(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
x with the values randomly ordered
}
\description{
Randomize a vector of values
}

