% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{readLogData}
\alias{readLogData}
\title{Read in a Aushon XML log file to get the deposition counts}
\usage{
readLogData(logFile)
}
\description{
Read in a Aushon XML log file to get the deposition counts
}

