% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{removeColumnNameSpaces}
\alias{removeColumnNameSpaces}
\title{Remove spaces in the column names}
\usage{
removeColumnNameSpaces(dt)
}
\arguments{
\item{dt}{a datatable}
}
\value{
the same datable with spaces in the column names removed.
}
\description{
\code{removeColumnNameSpaces} remove spaces in the colmun names of a datatable. This is useful to compare plates with slight differences in the metadata naming.
}

