% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{rotateMetadata}
\alias{rotateMetadata}
\title{Rotate the metadata 180 degrees in Array space}
\usage{
rotateMetadata(DT)
}
\arguments{
\item{DT}{A data.table of metadata with Spot, ArrayRow and ArrayColumn columns.}
}
\value{
The same data.table rotated 180 degrees in array space. The ArrayRow, Arraycolumn and Spot values are updated.
}
\description{
Rotate the metadata 180 degrees in Array space
}
