% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV_normalize_functions.R
\name{RUVArrayWithResiduals}
\alias{RUVArrayWithResiduals}
\title{Apply RUV normalization on a signal and its residuals}
\usage{
RUVArrayWithResiduals(k, Y, M, cIdx, signalName, verboseDisplay = FALSE)
}
\description{
Assumes there are signal values in the first half of each row
and residuals in the second half
}
