% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{RUVIII}
\alias{RUVIII}
\title{Perform RUV removal of unwanted variations
This function is written by Johann Gagnon-Bartsch and will become part of the
ruv package.}
\usage{
RUVIII(Y, M, ctl, k = NULL, eta = NULL, average = FALSE,
  fullalpha = NULL)
}
\description{
Perform RUV removal of unwanted variations
This function is written by Johann Gagnon-Bartsch and will become part of the
ruv package.
}

