% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{RUVIIIArrayWithResiduals}
\alias{RUVIIIArrayWithResiduals}
\title{Apply RUV3 normalization on a signal and its residuals}
\usage{
RUVIIIArrayWithResiduals(k, Y, M, cIdx, signalName)
}
\description{
Assumes there are signal values in the first half of each row
and residuals in the second half
}

