% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_levels.R
\name{se}
\alias{se}
\title{Calculate standard error of the mean}
\usage{
se(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the standard erro of the mean for x as a numeric value
}
\description{
Omit na values if present
}

