% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run_apps.R
\name{shiny_run_curate_units}
\alias{shiny_run_curate_units}
\title{Run the Curate Units Shiny Application}
\usage{
shiny_run_curate_units()
}
\description{
Usage:
}
\details{
1) Select a loaded dataset

   2) Unit --> Firing pattern
       - Select a unit from a list or small-multiples plot
       - hilight firing rate of unit in plots
            - Average firing rate of unit across treatments
            - Firing density across the experiment duration
            - Firing qq-plot across treatments

   3) Brushing Firing Patterns
       - Select query/response plots
       - Select unit in query plot
       - highlight unit in response plot
}
