% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{signalResidualMatrix}
\alias{signalResidualMatrix}
\title{Generate a matrix of signals and residuals}
\usage{
signalResidualMatrix(dt)
}
\arguments{
\item{dt}{a data.table with columns named BW and PrintSpot followed by one signal column}
}
\value{
A numeric matrix with BW rows and two sets of columns. The second
set of columns are the residuals from the medians of each column and have
"_Rel" appended to their names.
}
\description{
This function reorganizes a data.table into a matrix suitable for
RUV normalization.
}
