% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyLigandAnnotID}
\alias{simplifyLigandAnnotID}
\title{Create a short display version of the LigandAnnotID}
\usage{
simplifyLigandAnnotID(ligand, annotIDs)
}
\arguments{
\item{ligand}{The name of the ligand}

\item{annotIDs}{The annotIDs to be simplified}
}
\value{
a vector with the ligand and the last string of each annotID pasted with a _
}
\description{
Create a short display version of the LigandAnnotID
}

