% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{spotCellDensitiesScanR}
\alias{spotCellDensitiesScanR}
\title{Calculate the neighborhood density around each cell}
\usage{
spotCellDensitiesScanR(spot, radius = (max(spot$X) - min(spot$X))/5)
}
\arguments{
\item{spot}{A datatable with X and Y columns}

\item{radius}{The radial distance that defines the neighborhood around the cell}
}
\value{
A numeric vector of length nrow(spot) with the cell density values
}
\description{
Calculate the neighborhood density around each cell
}
