% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{spotNormIntensities}
\alias{spotNormIntensities}
\title{Add spot level normalizations for median intensities}
\usage{
spotNormIntensities(dt)
}
\description{
Add spot level normalizations for median intensities
}
