% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-waveform_correlation_matrix.R
\name{waveform_correlation_matrix}
\alias{waveform_correlation_matrix}
\title{plot Correlation matrix of waveforms}
\usage{
waveform_correlation_matrix(waveform, experiment_tag, plot_width = 10,
  plot_height = 10, output_base = "product/figures", verbose = TRUE)
}
\description{
Plot a correlation matrix as a heatmap between all waveforms in the experiment
}
\details{
waveform:
  data.frame with columns [neuron_index, timestamp, voltage]

experiment_tag:
  identifier for the experiment used in the figure subtitle and output filename

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/figures/waveform_correlation_matrix_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
