% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-waveform_lattice.R
\name{waveform_lattice}
\alias{waveform_lattice}
\title{Simple lattice plot of the waveforms}
\usage{
waveform_lattice(waveform, experiment_tag, plot_width = 10,
  plot_height = 10, output_base = "product/figures", verbose = TRUE)
}
\description{
Plot a grid of figures, with x-axis in microseconds and y-axis in voltage
  One for each waveform in the waveform data.frame
}
\details{
waveform:
  data.frame with columns [neuron_index, timestamp, voltage]

experiment_tag:
  identifier for the experiment used in the figure subtitle and output filename

plot_width/plot_height:
  dimensions of the output plot

returns:
  the ggplot2
  Saves the result to product/figures/waveform_lattice_<experiment_tag>_<date_code>.(pdf|png)
  It save both .pdf and .png because it's easier to email etc small pngs
  while for use in an a manuscript having the vector version means that it can be tweaked with illustrator
}
