% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{wellAN}
\alias{wellAN}
\title{Returns a character vector of alphanumeric well names}
\usage{
wellAN(nrRows, nrCols)
}
\arguments{
\item{nrRows}{The number of rows in the plate or array.}

\item{nrCols}{The number of columns in the plate or array.}
}
\value{
A character vector of the well names
}
\description{
\code{wellAN} is a convenience function that provides alphanumeric names
for well plates or arrays.
}

