% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{wellLevelData}
\alias{wellLevelData}
\title{Summarize the cell level data to the well level}
\usage{
wellLevelData(cd)
}
\description{
Summarize the cell level data to the well level
}

