% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{writeCellLevel}
\alias{writeCellLevel}
\title{Write the cell level raw data and metadata Level 1 data.}
\usage{
writeCellLevel(dt, path, barcode, verbose = FALSE)
}
\description{
Write the cell level raw data and metadata Level 1 data.
}
