% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{calcTheta}
\alias{calcTheta}
\title{Calculate the polar coordinate theta value}
\usage{
calcTheta(x, y)
}
\arguments{
\item{x}{The x numeric value in cartesian coordinates.}

\item{y}{The y numeric value in cartesian coordinates.}
}
\value{
The theta value of the X and y converted to polar coordinates.
}
\description{
Return the polar coordinate theta value that starts at 0 along the positive Y axis and increases to 360 in a clockwise direction.
}
