% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{cleanColumnNames}
\alias{cleanColumnNames}
\title{Change the periods in the column names to spaces}
\usage{
cleanColumnNames(dt)
}
\description{
Change the periods in the column names to spaces
}

