% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_levels.R
\name{createl3}
\alias{createl3}
\title{Summarize cell level data to the spot level}
\usage{
createl3(cDT, lthresh = lthresh, seNames = NULL)
}
\arguments{
\item{cDT}{The datatable of cell level data to be summarized}

\item{lthresh}{The threshold used in the loess model to define low cell count
regions
@return A datatable of spot-level, median summarized data with standard error values and 
metadata}
}
\description{
Median summarize the cell level normalized values and the most biologically 
interpretable raw data at each spot, calculate the standard errors and add
SE columns for all median summarized data
}

