% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{createRUVM}
\alias{createRUVM}
\title{Create an M matrix for the RUV normalization}
\usage{
createRUVM(dt, replicateCols = c("CellLine", "Ligand", "Drug"))
}
\arguments{
\item{dt}{The datatable to be normalized that has columns Barcode, Well and 
those in the replicateCols parameter.}

\item{replicateCols}{A character vector of column names in dt that define a replicate well.}
}
\value{
A matrix suitable for defining the dataset structure in RUV normalization
}
\description{
Each row is for a unit to be normalized and
each column is a unique replicate type
Each row will have a 1 to indicate the replicate type
all other values will be 0
}
