% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRUVM.R, R/normalize_functions.R
\name{createRUVM}
\alias{createRUVM}
\title{Create an M matrix for the RUV normalization}
\usage{
createRUVM(dt)

createRUVM(dt)
}
\arguments{
\item{dt}{The datatable to be normalized. There must be a BWL column
that defines the unique units, a Ligand column that defines the unique
wells and a SignalType column where a value of "Signal" denotes which 
BWL and Ligand rows should be included in the M matrix.}
}
\value{
A datatable with values of 1 and 0.
}
\description{
The M matrix holds the structure of the dataset in RUV normalization.
There is one row for each unit to be normalized and
one column for each unique unit type
Each row will have one 1 to indicate the unit type
all other values will be 0

Create an M matrix for the RUV normalization
that has one row for each unit to be normalized and
one column for each unique unit type
Each row will have a 1 to indicate the unit type
all other values will be 0
}

