% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_functions.R
\name{CV}
\alias{CV}
\title{Calculate Coefficient of Variation (CV)}
\usage{
CV(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The CV of the non-na values in the x.

 @export
}
\description{
Calculates the CV of a numeric vector as the standard deviation over the mean. All NA values are removed from the input vector.
}
