% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getLogMetadata}
\alias{getLogMetadata}
\title{Get deposition values from an Aushon log dta files}
\usage{
getLogMetadata(fn)
}
\arguments{
\item{fn}{full path and file name for the log metadata file}
}
\value{
a datatable with an integer Depostion column
}
\description{
Get deposition values from an Aushon log dta files
}
