% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSpotLevelData}
\alias{getSpotLevelData}
\title{Get the spot level data for a study}
\usage{
getSpotLevelData(studyName, path)
}
\arguments{
\item{studyName}{The name of a study that contains data from one or more plates}

\item{path}{The path to the directory that contains the barcode level subdirectories}
}
\value{
A datatable with the annotated data for all plates in the study. Any data for fiducials and 
blank spots is filtered out.
}
\description{
Get the spot level data for a study
}
