% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getSpotMetadata}
\alias{getSpotMetadata}
\title{Read the spot metadata from a gal file}
\usage{
getSpotMetadata(fn)
}
\arguments{
\item{fn}{full path and file name for the spot metadata file}
}
\value{
a datatable with the spot level metadata
}
\description{
Read the spot metadata from a gal file
}
