% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessCellFunctions.R
\name{getWellMetadata}
\alias{getWellMetadata}
\title{Read the well metadata from a multi-sheet Excel file}
\usage{
getWellMetadata(path, barcode)
}
\arguments{
\item{path}{path to the directory with barcode level subdirectories}

\item{barcode}{barcode or plate ID}
}
\value{
a datatable with the well level metadata
}
\description{
Read the well metadata from a multi-sheet Excel file
}
