% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{kmeansCluster}
\alias{kmeansCluster}
\title{kMeans cluster an intensity value}
\usage{
kmeansCluster(x, value, ctrlLigand = "HighSerum")
}
\arguments{
\item{x}{A dataframe or data.table}

\item{value}{A character vector of one column name in x that conatins non-negative intensity values}

\item{ctrlLigand}{A character vector of one Ligand name in x}
}
\value{
Cluster assignments of length nrow(x)for the value column based on a kmeans threshold calculated from the ligand values. The values are 1 for intensities above the threshold and 0 for values equal to or less than the trheshold.
}
\description{
Returns cluster assignments for intensity values based on a control ligand
}
