% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{labelOuterCells}
\alias{labelOuterCells}
\title{Classify Outer cells}
\usage{
labelOuterCells(x, thresh = 0.75)
}
\arguments{
\item{x}{A numeric vector of RadialPositions}

\item{thresh}{A quantile value between 0 and 1 used to threshold x. The returned logical will be TRUE for cells with x values in quantiles greater than thresh.}
}
\value{
A logical vector the length of x with a TRUE value for the Outer cells
}
\description{
\code{labelOuterCells} Determine if a cell is an Outer cell in the spot
}
