% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{medianNorm}
\alias{medianNorm}
\title{Median normalize a vector}
\usage{
medianNorm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A numeric vector that is \code{x} divided by its median
}
\description{
Median normalize a vector
}
