% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{mergeMetadata}
\alias{mergeMetadata}
\title{merge the data and metadata on the well index column}
\usage{
mergeMetadata(dt, mdf)
}
\arguments{
\item{dt}{A datatable of data values}

\item{mdf}{A datatable of metadata}
}
\value{
A datatable of annotated data that is merged on the Well values.
}
\description{
\code{mergeMetadata} merges the well metadata with data on the well value.
}

