% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{mergeSpotMetadata}
\alias{mergeSpotMetadata}
\title{merge the data and metadata on the well index column}
\usage{
mergeSpotMetadata(dt, mdf)
}
\description{
merge the data and metadata on the well index column
}

