% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUV_normalize_functions.R
\name{normRUVLoessResiduals}
\alias{normRUVLoessResiduals}
\title{Apply RUV and Loess Normalization to the signals in a MEMA dataset}
\usage{
normRUVLoessResiduals(dt, k)
}
\arguments{
\item{dt}{A datatable with data and metadata to be normalzed. There must be CellLine, Barcode, Well, Ligand, Drug and ECMp metadata columns}

\item{k}{The number of factors to be reomved in the RUV normalization}
}
\description{
Apply RUV and Loess Normalization to the signals in a MEMA dataset
}
