% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_functions.R
\name{readLogData}
\alias{readLogData}
\title{Read in and parse an Aushon XML log file}
\usage{
readLogData(logFile)
}
\arguments{
\item{logFile}{An Aushon logfile}
}
\value{
A datatable keyed by Row and Column with Depositions and
 PrintOrder columns.
}
\description{
Read in and parse an Aushon XML log file
}
