% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_functions.R
\name{signalResidualMatrix}
\alias{signalResidualMatrix}
\title{Generate a matrix of signals and residuals}
\usage{
signalResidualMatrix(dt)
}
\arguments{
\item{dt}{a data.table with columns named BWL, SERC, SignalType and a column 
named by the unique value in the SignalType column.}
}
\value{
A numeric matrix with BWL rows and two sets of SERC columns. The second
set of SERC columns are the residuals from the medians of each column and have
"_Rel" appended to their names.
}
\description{
This function reorganizes a data.table into a matrix suitable for
RUV3 normalization.
}

