% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derivedParameter_functions.R
\name{spotCellDensities}
\alias{spotCellDensities}
\title{Calculate the neghborhood density around each cell}
\usage{
spotCellDensities(spot, radius = (max(spot$X) - min(spot$X))/5)
}
\arguments{
\item{spot}{A datatable with X and Y columns}

\item{radius}{The radial distance that defines the neighborhood around the cell}
}
\value{
A numeric vector of length nrow(spot) with the cell density values
}
\description{
Calculate the neghborhood density around each cell
}

