% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMA_Test_Exp_Functions.R
\name{summarizeToSpot}
\alias{summarizeToSpot}
\title{Summarize cell data to spot level}
\usage{
summarizeToSpot(cd)
}
\arguments{
\item{cd}{A datatable of cell level data to be summarized.}
}
\value{
A datatable of the spot level data.
}
\description{
\code{summarizeToSpot} summarizes cell level data to the spot level.
}

